requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# HTTP client (Stage 1 collector)
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.4

# ML / Data
# TensorFlow 2.13+ ships native arm64 wheels — works on Apple Silicon with Python 3.11
tensorflow>=2.13.0
scikit-learn>=1.4.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0

# Serialisation
pydantic>=2.6.0

# Config loading
pyyaml>=6.0.1

# Persistence
joblib>=1.3.0
